feat(extract detail_game_info/ extract enum from player_info.h)
