+++ includes/detail_game_info.h
@@ -1,0 +1,6 @@
+#ifndef _DETAIL_GAME_INFO_H_
+#define _DETAIL_GAME_INFO_H_
+
+void detail_game_info();
+
+#endif  /* !_DETAIL_GAME_INFO_H_ */
